wip: add workflow history page
