+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/index.svelte
@@ -1,26 +1,5 @@
 <script lang="ts">
-  import { onMount } from 'svelte';
-  import { page } from '$app/stores';
-  import { goto } from '$app/navigation';
-
-  import { eventViewType, eventSortOrder } from '$lib/stores/event-view';
-  import { routeForEventHistory } from '$lib/utilities/route-for';
-
-  const { namespace, workflow, run } = $page.params;
-
-  onMount(async () => {
-    const queryParams = { sort: $eventSortOrder };
-    goto(
-      routeForEventHistory({
-        view: $eventViewType,
-        queryParams,
-        namespace,
-        workflow,
-        run,
-      }),
-      {
-        replaceState: true,
-      },
-    );
-  });
+  import WorkflowHistory from '$lib/pages/workflow-history.svelte';
 </script>
+
+<WorkflowHistory />
